requirements.txt
kivy
google-generativeai 
pytest 
rapidfuzz 